Modulo4NoSqlBD/mongodb/tareas/Challenge/requirements.txt

pymongo
requests
python-dotenv